Add a smart `rs` alias that also configures Thin.
The new `rs` command takes 2 parameters, a command to send to Thin and
the port for the server to run on (if the command is 'start'). I use
Thin proxied to Apache for all of my development server needs, so this
makes it easier to start and stop my app servers on demand.

The `thin restart` command doesn't seem to work. Looking into it..