Add vendor/gems/bin to the PATH globally.
Provides RVM-like magic for selecting the proper binary directories
and GEM_HOMEs for vendored gems. We now vendor gems using Bundler
instead of using RVM to manage gemsets, and if we need to in the future
we can use rbenv to manage different versions of Ruby. All of this
is glued together with Autoenv, which runs .env files in each dir,
and can be used for selecting the right Ruby as well as modifying
other environment-level configuration settings on a per-project
basis.