Save code to replace text in files within a dir.
This may be turned into a function/command someday. Hypothetically,
it would take two arguments: the sed expression that's to be run
on every file in a glob, and the glob pattern to then invoke on
the function. It should infer that the first argument of this sed
expression is meant to be input into the `grep` call that will
return the contents of each file in the glob.