Add fn for wrapping global gems
This is kind of a hacky solution to an ongoing Ruby problem. When
you want to use a gem's binary in any directory, regardless of
what GEM_HOME is set to, you can wrap the command with `globalize`
to execute that command in the context of the $GLOBAL_GEM_HOME,
as set in ~/.env. This is similar to how `bundle exec` works, by
executing the succeeding command in the context of the isolated
GEM_HOME. However, in the event that we have GEM_HOME overridden
in a local .env, `globalize` can ostensibly reverse this tactic
and execute the gem in the global gemset.