Use CtrlP instead of Command-T
While Command-T is considerably faster than CtrlP, especially on
startup, it requires that you have the same version of Ruby running that
you compiled it against, and since we can no longer be sure of what
version of Ruby will be running at the time Command-T is being invoked,
we can't rely on it to always be working. Therefore, CtrlP will take its
place and require the use of a Git repo in order to list out all files.