Configure GPG to send secure emails with Mutt.
This is the main thing I was trying to accomplish with Mutt, the ability
to send PGP-signed and/or encrypted emails. To do this, I'm using GPG
and sourcing Mutt's default GPG configuration, and configuring Mutt to
automatically sign messages that I send, as well as automatically decrypt
messages sent to me which have come in encrypted. I've also added a
post-recieve hook for offlineimap that will help with updating the
search index for locally-stored mail messages. Had to install MSMTP in
order to get messages sent for some reason.